@@ -5,12 +5,12 @@ version: 1.0.0
 author: Germán Alonso <galonsos@gmail.com>
 
 environment:
-  sdk: '>=2.1.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  url_launcher: ^5.0.2
+  url_launcher: ^6.0.2
 
 dev_dependencies:
-  test: ^1.0.0
+  test: ^1.16.5
